# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Plexus API
  description: Enterprise project management API
  version: 1.0.0
  contact:
    name: Plexus
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.plexus.app/api/v1
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
      required: [access_token, refresh_token, expires_in]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        avatar_url:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, user]
        created_at:
          type: string
          format: date-time
      required: [id, email, display_name, role, created_at]
...
  parameters:
    orgSlug:
      name: orgSlug
      in: path
      required: true
      schema:
        type: string
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required: [refresh_token]
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
...
  /me:
    get:
      tags: [Users]
      operationId: getMe
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags: [Users]
      operationId: updateMe
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                avatar_url:
                  type: string
      responses:
        '204':
          description: Updated
...
  /orgs/{orgSlug}/api-keys:
    get:
      tags: [API Keys]
      operationId: listAPIKeys
      parameters:
        - $ref: '#/components/parameters/orgSlug'
      responses:
        '200':
          description: Org API keys (secrets redacted)
    post:
      tags: [API Keys]
      operationId: createAPIKey
      parameters:
        - $ref: '#/components/parameters/orgSlug'
      responses:
        '201':
          description: Created (returns raw key once)
  /orgs/{orgSlug}/audit:
    get:
      tags: [Audit]
      operationId: listAuditEvents
      parameters:
        - $ref: '#/components/parameters/orgSlug'
      responses:
        '200':
          description: Audit events
...